Audit/hardening pass: fix clean-build break, ring-3 syscall security, EFER.SCE, better panic output
- sha256.h: replace #include <stdint.h> with kernel.h (broke clean build)
- Fix all pointer-truncation warnings (uintptr_t cast throughout
  kernel.c, vfs.c, vbe.c, sb16.c, rtl8139.c, initramfs.c, smp.c)
- keyboard.c: trim over-initializers from keymap tables (C auto-zero-fills)
- ext2.c, rtc.c, compositor.c, apic.c: remove dead variables/code
- isr.c/isr_stubs.asm: panic now prints faulting RIP, CS, ring, error code
- syscall.c: user pointer validation (user_ptr_ok/user_str_ok) against
  canonical user half [0x1000, 0x800000000000)
- syscall.c: fd translation table (small int ufd -> internal VFS handle)
  prevents kernel VFS pointer forgery from ring 3
- syscall.c: enable EFER.SCE in setup_syscall_msrs — was missing, causing
  #UD on syscall in ring 3 (root cause of user-process crash-on-entry)
- AGENTS.md: v5.4.0 entry, ring-3 syscall fix plan, security notes
- README.md: updated known issues for ring-3 and scheduling status

+++ kernel/kernel.c
@@ -988,7 +988,7 @@ static void cmd_hexdump(int argc, char** argv) {
         count = 1024;
     }
     printf("Dumping %d bytes from 0x%x:\n", count, addr);
-    uint8_t* ptr = (uint8_t*)addr;
+    uint8_t* ptr = (uint8_t*)(uintptr_t)addr;
     for (uint32_t i = 0; i < count; i += 16) {
         char addr_str[9];
         uint32_t a = addr + i;
@@ -1899,20 +1899,20 @@ void init_load_modules(void) {
     uint32_t mods_count = mb[5];
     uint32_t mods_addr = mb[6];
     printf("[MODULES] %d module(s) at %x\n", mods_count, mods_addr);
-    uint32_t* mod_entry = (uint32_t*)mods_addr;
+    uint32_t* mod_entry = (uint32_t*)(uintptr_t)mods_addr;
     for (uint32_t i = 0; i < mods_count; i++) {
         uint32_t mod_start = mod_entry[0];
         uint32_t mod_end = mod_entry[1];
         uint32_t mod_str = mod_entry[2];
         uint32_t mod_size = mod_end - mod_start;
-        const char* name = (const char*)mod_str;
+        const char* name = (const char*)(uintptr_t)mod_str;
         char path[64];
         if (name && *name) {
             snprintf(path, sizeof(path), "/%s", name);
         } else {
             snprintf(path, sizeof(path), "/boot/module%d", i);
         }
-        vfs_create_from_mem(path, (uint8_t*)(uint32_t)mod_start, mod_size);
+        vfs_create_from_mem(path, (uint8_t*)(uintptr_t)mod_start, mod_size);
         mod_entry += 4;
     }
 }
